Enhance profile image upload functionality; refactor form handling and integrate file submission in Profile component
